requirements.txt
Django==1.6.5
South==0.8.4
Unipath==1.0
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.5
django-mptt==0.6.1
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.3
pystache==0.5.4
requests==2.3.0
static==1.0.2
wsgiref==0.1.2
